[BOSK-28] feat: Add blocks to the main package. Fixed examples using blocks from the package
