+++ bosk/block/zoo/data_weighting/__init__.py
@@ -1,0 +1,1 @@
+from .data_weighting import WeightsBlock
