+++ bosk/block/zoo/input_plugs/__init__.py
@@ -1,0 +1,1 @@
+from .input_plugs import InputBlock, TargetInputBlock
