+++ bosk/block/zoo/metrics/__init__.py
@@ -1,0 +1,1 @@
+from .metrics import AccuracyBlock, RocAucBlock, F1ScoreBlock
