+++ bosk/block/zoo/models/classification/__init__.py
@@ -1,0 +1,1 @@
+from .classification_models import RFCBlock, ETCBlock
