+++ bosk/block/zoo/models/regression/__init__.py
@@ -1,0 +1,1 @@
+from .regression_models import RFRBlock, ETRBlock
