+++ bosk/block/zoo/preprocessing/__init__.py
@@ -1,0 +1,1 @@
+import sklearn.preprocessing as preprocessing
