+++ requirements.txt
@@ -1,0 +1,7 @@
+-e git+https://github.com/NTAILab/bosk@78269952e9bba3c0b2c43d7c28ecfd5d59a82ec1#egg=bosk
+joblib==1.1.0
+numpy==1.23.2
+scikit-learn==1.1.2
+scipy==1.9.1
+sklearn==0.0
+threadpoolctl==3.1.0
