--- requirements_dev.txt
@@ -1,1 +1,0 @@
-.
